// src/model/interfaces/docente/listaCarta.ts
export default interface ListaCartaEmpresaDocente {
    cartaId: number
    empresaNombre: string
    empresaRuc: string
    empresaId: number
    direccion: string
    depProvDist: string
    representante: string
    repDni: string
    cargoNombre: string
    gradoNombre: string
    fechaRegistro: string
    estadoCarta: number
}